test ngrx pour affichage de notifications centralisées

+++ todolist/src/app/modules/explorer/todoCommands/todoCommands.component.ts
@@ -5,6 +5,12 @@
 import {Component, EventEmitter, Input, Output} from '@angular/core';
 import {Todo} from "../../../model/Todo";
 import {TodoService} from "../../../../services/todo.service";
+import {MatDialog} from "@angular/material/dialog";
+import {ConfirmDialogComponent} from "../confirmDialog/confirmDialog.component";
+import {MatSnackBar} from "@angular/material/snack-bar";
+import {Store} from "@ngrx/store";
+import {setNotificationList} from "../../../store/notification.actions";
+import {Notification} from "../../../model/Notification";
 
 @Component({
   selector: 'todoCommands',
@@ -18,13 +24,58 @@ export class TodoCommandsComponent {
   @Input() fromList: boolean = false;
   @Output('delete') readonly delete = new EventEmitter<Todo>();
 
-  constructor(private todoService: TodoService) {
+  constructor(private todoService: TodoService,
+              private dialog: MatDialog,
+              private snackbar: MatSnackBar,
+              private store: Store<{ notification: string }>) {
   }
 
   onDelete(): void {
 
-    this.todoService.deleteTodo(this.todo)
-      .subscribe(() => this.delete.emit(this.todo))
+    const dialogRef = this.dialog.open(ConfirmDialogComponent,
+      {
+        width: '300px',
+        data: {title: "Delete the task", message: `Are you sure you want to delete the task : ${this.todo?.title}`}
+      })
+
+    dialogRef.afterClosed().subscribe(confirm => {
+      if (confirm) {
+        this.todoService.deleteTodo(this.todo)
+          .subscribe({
+            next: () => {
+              let notifications: Array<Notification> = [{
+                severity: 'success',
+                summary: 'Task delete',
+                detail: `The task "${this.todo.title}" has been successfully deleted ! `
+              }];
+
+              this.store.dispatch(setNotificationList({notifications}))
+              //this.snackbar.open("The task has been successfully deleted ! ", "Keep going", {duration: 5000})
+              this.delete.emit(this.todo);
+            },
+            error: (err) => {
+              console.log("error")
+              console.log(err)
+
+              let notifications: Array<Notification> = [{
+                severity: 'error',
+                summary: 'Task delete error',
+                detail: `An error occured when deleting the task "${this.todo.title}" ! `
+              }];
+
+              this.store.dispatch(setNotificationList({notifications}))
+            },
+            complete: () => {
+              console.log("complete")
+            }
+          })
+      }
+    })
+
   }
 
+  onDone() {
+    this.todoService.switchDone(this.todo)
+      .subscribe((done) => this.todo.done = done);
+  }
 }
